+++ CHANGELOG.md
@@ -26,8 +26,12 @@ The changelog for `MessageKit`. Also see the [releases](https://github.com/Messa
 - `scrollsToBottomOnFirstLayout` property to automatically scroll to the bottom of `MessagesCollectionView` on first load.
 [#213](https://github.com/MessageKit/MessageKit/pull/213) by [@FraDeliro](https://github.com/FraDeliro).
 
+- `scrollsToBottomOnKeyboardDidBeginEditing` property to automatically scroll to the bottom of `MessagesCollectionView` when the keyboard begins editing.
+[#217](https://github.com/MessageKit/MessageKit/pull/217) by [@SD10](https://github.com/SD10).
+
 - `additionalTopContentInset` property to `MessagesColectionViewController` to allow users to account for extra subviews.
 [#218](https://github.com/MessageKit/MessageKit/pull/218) by [@SD10](https://github.com/SD10).
+
 
 ### Fixed